+++ leda/components/DropDownSelect/handlers.ts
@@ -235,16 +235,15 @@ export const createKeyDownHandler = ({
 
     if (isOpen) mergeState({ isOpen: false });
 
-    if (fullData.length === 0) return;
-
-    const value = getText(highlightedSuggestion, textField);
+    const textValue = getText(highlightedSuggestion, textField); // onFilterChange require empty string as value. null value can't be used there
+    const value = textValue || null; // Convert empty string to null. '' -> null
 
     if (isFunction(onFilterChange)) {
       const customEvent = {
         ...ev,
         component: {
           name,
-          value,
+          value: textValue,
           suggestion: highlightedSuggestion,
         },
       };