+++ leda/index.ts
@@ -15,6 +15,7 @@ import {
   useInterval,
   useProps,
   useValue,
+  useRunAfterUpdate,
 } from './utils';
 
 import { A } from './components/A';
@@ -132,6 +133,7 @@ import * as VStepperTypes from './components/VStepper/types';
 import * as commonTypes from './commonTypes';
 
 import { form } from './form';
+import { maskValue } from './src/MaskedInputBase/helpers';
 
 const utils = {
   bindFunctionalRef,
@@ -149,6 +151,8 @@ const utils = {
   useInterval,
   useProps,
   useValue,
+  useRunAfterUpdate,
+  maskValue,
 };
 
 export {