+++ leda/src/DateTimeInput/helpers.ts
@@ -5,9 +5,9 @@ import {
 import { Values } from '../../commonTypes';
 import { getClassNames, getIsEmptyAndRequired } from '../../utils';
 import { GlobalDefaultTheme } from '../../utils/useTheme';
-import { COMPONENT_TYPES } from './constants';
+import { COMPONENT_TYPES, HOURS_LIMITS, MINUTES_LIMITS } from './constants';
 import {
-  DateWithToDateMethod, DateTimeInputProps, DateTimeInputState, NormalizeValueArgs,
+  DateWithToDateMethod, DateTimeInputProps, DateTimeInputState, NormalizeValueArgs, TimeLimits,
 } from './types';
 
 // извлекает число по паттерну и формату. Пример: ("dd.MM.yy", "dd", "18.05.19") => 18
@@ -163,6 +163,20 @@ export const getInputWrapperClassNames = (theme: GlobalDefaultTheme['dateTimeInp
     { [theme.inputWrapperRequired]: getIsEmptyAndRequired(value, props.isRequired) },
     { [theme.wrapperDisabled]: props.isDisabled },
   );
+};
+
+const normilizeNumber = (value: number, rules: TimeLimits): number => {
+  const [min, max] = rules;
+  if (value < min) return min;
+  if (value > max) return max;
+  return value;
+};
+
+/* Нормализуем ограничители, приводим к минимальному или максимальному значению, или оставляем как есть */
+export const normalizeTimeLimits = (timeLimits: TimeLimits | undefined): TimeLimits | undefined => {
+  if (!timeLimits) return undefined;
+  const [hours, minutes] = timeLimits;
+  return [normilizeNumber(hours, HOURS_LIMITS as TimeLimits), normilizeNumber(minutes, MINUTES_LIMITS as TimeLimits)];
 };
 
 export const convertToDate = (dateValue?: DateWithToDateMethod): DateWithToDateMethod | undefined => {